refactor(server): split WebSocket frame handlers out of the session loop
Move binary/configure/stop/flush handlers into ws/frames.rs so
handle_ws_inner stays an orchestration loop. Protocol and keepalive
behavior are unchanged.

Constraint: no WS protocol or drain-path change
Rejected: further split of greedy_decode_impl | already has select_token/commit_non_blank
Confidence: high
Scope-risk: moderate
